""""welcome"""" message

+++ index.mjs
@@ -3,20 +3,26 @@ dotenv.config();
 
 import Discord from "discord.js";
 import fs from "fs";
-const db = "messages.txt";
 const client = new Discord.Client({ intents: ["GUILDS", "GUILD_MESSAGES"] });
 
 // init message db
-if(!fs.existsSync(db)) fs.writeFileSync(db, "");
-const stream = fs.createWriteStream(db, { flags: "a" });
-const messages = new Set(fs.readFileSync(db, "utf8").split("\x00"));
+const messageDb = "messages.txt";
+if(!fs.existsSync(messageDb)) fs.writeFileSync(messageDb, "");
+const stream = fs.createWriteStream(messageDb, { flags: "a" });
+const messages = new Set(fs.readFileSync(messageDb, "utf8").split("\x00"));
+
+// init "welcome" message
+const welcomeDb = "welcomes.txt";
+if(!fs.existsSync(welcomeDb)) fs.writeFileSync(welcomeDb, "");
+const welcomes = new Set(fs.readFileSync(welcomeDb, "utf8").split("\x00"));
+const welcomeMsg = "warning: do not send the same message twice. you will be muted.";
 
 // init mutes
-const mutefile = "mutes.txt";
+const muteDb = "mutes.txt";
 const mutes = new Map();
 const muted = "894095301042901063";
 
-// init reply to reaction
+// init reply => reaction
 const reactions = new Map();
 reactions.set("¯\\_(ツ)_/¯", "🤷");
 reactions.set("oh no", "894300904235352124");
@@ -43,11 +49,15 @@ client.on("ready", () => {
 });
 
 client.on("messageCreate", async (msg) => {
+	// ignore system messages (no content)
+	if(msg.system) return;
 	if(!msg.content) {
+		// if there are attachments, its not a sys
 		if(msg.attachments.size) await msg.delete();
 		return;
 	}
 
+	// convert messages into emoji
 	for(let [trigger, emoji] of reactions) {
 		if(msg.content !== trigger) continue;
 		let ref = msg.reference ?
@@ -57,7 +67,8 @@ client.on("messageCreate", async (msg) => {
 		msg.delete();
 		return;
 	}
-	
+
+	// the core code
 	const content = msg.content?.trim().toLowerCase().replace(/[^a-z ]/g, "");
 	if(messages.has(content)) {
 		msg.delete();
@@ -72,6 +83,13 @@ client.on("messageCreate", async (msg) => {
 		if(!mutes.has(id)) updateMutes();
 		msg.member.roles.add(muted);
 		mutes.set(id, { length, timeout });
+
+		if(!welcomes.has(msg.author.id)) {
+			const chan = await msg.author.createDM();
+			await chan?.send(welcomeMsg);
+			welcomes.add(msg.author.id);
+		}
+
 		return;
 	}
 
@@ -86,7 +104,11 @@ setInterval(() => {
 }, 1000 * 60 * 30);
 
 function updateMutes() {
-	fs.writeFileSync(mutefile, [...mutes.keys()].join(","));
+	fs.writeFileSync(muteDb, [...mutes.keys()].join(","));
+}
+
+function updateWelcomes() {
+	fs.writeFileSync(welcomeDb, [...welcome].join("\x00"));
 }
 
 async function getBefore(msg) {
